Renamed getPrice to getPrices helper
